update: add input validation and logger

+++ src/CollabDocumentEditor.Web/Controllers/DocumentsController.cs
@@ -1,3 +1,4 @@
+using CollabDocumentEditor.Core.Common;
 using CollabDocumentEditor.Core.Dtos;
 using CollabDocumentEditor.Core.Interfaces.Services;
 using Microsoft.AspNetCore.Authorization;
@@ -11,16 +12,23 @@ namespace CollabDocumentEditor.Web.Controllers;
 public class DocumentsController : ControllerBase
 {
     private readonly IDocumentService _documentService;
+    private readonly ILogger<DocumentsController> _logger;
     
-    public DocumentsController(IDocumentService documentService)
+    public DocumentsController(IDocumentService documentService, ILogger<DocumentsController> logger)
     {
         _documentService = documentService ?? throw new ArgumentNullException(nameof(documentService));
+        _logger = logger ?? throw new ArgumentNullException(nameof(logger));
     }
     
     [HttpGet("{documentId}", Name = "GetDocument")]
     [Authorize(Policy = "DocumentRole")]
     public async Task<IActionResult> GetDocumentAsync(Guid documentId)
     {
+        if (documentId == Guid.Empty)
+        {
+            return BadRequest("Document id cannot be empty");
+        }
+        
         var document = await _documentService.GetDocumentAsync(documentId);
         return Ok(document);
     }
@@ -29,6 +37,11 @@ public class DocumentsController : ControllerBase
     [Authorize(Policy="DocumentRole")]
     public async Task<IActionResult> GetUserDocumentsAsync(Guid userId)
     {
+        if (userId == Guid.Empty)
+        {
+            return BadRequest("User id cannot be empty");
+        }
+        
         var documents = await _documentService.GetUserDocumentsAsync(userId);
         return Ok(documents);
     }
@@ -38,6 +51,9 @@ public class DocumentsController : ControllerBase
     public async Task<IActionResult> CreateDocumentAsync([FromBody] CreateDocumentDto dto)
     {
         var document = await _documentService.CreateDocumentAsync(dto);
+        
+        _logger.LogInformation("Document created: {document}", document);
+        
         return CreatedAtRoute(
             routeName: "GetDocument",
             routeValues: new { documentId = document.Value.Id },
@@ -49,12 +65,20 @@ public class DocumentsController : ControllerBase
     [Authorize(Policy="DocumentRole")]
     public async Task<IActionResult> UpdateDocumentAsync(Guid documentId, [FromBody] UpdateDocumentDto dto)
     {
+        if (documentId == Guid.Empty)
+        {
+            return BadRequest("Document id cannot be empty");
+        }
+        
         if (documentId != dto.Id)
         {
+            _logger.LogWarning("Document ID mismatch. URL ID: {UrlId}, DTO ID: {DtoId}", documentId, dto.Id);
             return BadRequest("Document id in the request body does not match the document id in the URL");
         }
         
         var document = await _documentService.UpdateDocumentAsync(dto);
+        
+        _logger.LogInformation("Document updated: {document}", document);
         return Ok(document);
     }
     
@@ -62,7 +86,14 @@ public class DocumentsController : ControllerBase
     [Authorize(Policy="DocumentRole")]
     public async Task<IActionResult> DeleteDocumentAsync(Guid documentId)
     {
+        if(documentId == Guid.Empty)
+        {
+            return BadRequest("Document id cannot be empty");
+        }
+        
         await _documentService.DeleteDocumentAsync(documentId);
+        
+        _logger.LogInformation("Document deleted: {documentId}", documentId);
         return NoContent();
     }
     
@@ -71,6 +102,8 @@ public class DocumentsController : ControllerBase
     public async Task<IActionResult> ShareDocumentAsync([FromBody] ShareDocumentDto dto)
     {
         await _documentService.ShareDocumentAsync(dto);
+        
+        _logger.LogInformation("Document shared: {documentId}", dto.DocumentId);
         return NoContent();
     }
 }
